pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "flask-cms"
authors = [
    {name = "Frédérick NEY", email = "frederick.ney@gmail.com"},
]
description = "Flask CMS"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
classifiers = [
"Environment :: Web Environment",
"Intended Audience :: Developers",
"Operating System :: OS Independent",
"Framework :: Flask",
"Topic :: Software Development :: Libraries :: Python Modules",
'Programming Language :: Python :: 3.7',
'Programming Language :: Python :: 3.8',
'Programming Language :: Python :: 3.9',
'Programming Language :: Python :: 3.10',
'Programming Language :: Python :: 3.11',
'Programming Language :: Python :: 3.12',
'Programming Language :: Python :: 3.13',
]
version="0.0.1"
dependencies = [
    "flask-login-saml",
    "flask-login-oidc",
    "flask-framework-mvc",
    'importlib-metadata; python_version>"3.7"'
]

[options.package_data]
module = ["flask_cms/template/**/*.html", "flask_cms/template/**/*.*"]

[tool.setuptools.packages.find]
exclude = ["models*", "server*", "config*", "log*", "uploads*", "extensions*", "static*", "template*", "controllers*"]

[project.urls]
Homepage = "https://github.com/frederickney/flask-cms"
Documentation = "https://github.com/frederickney/flask-cms/blob/master/README.md"
Repository = "https://github.com/frederickney/flask-cms"
Issues = "https://github.com/frederickney/flask-cms/issues"
source = "https://github.com/frederickney/flask-cms"

